requirements.txt
Flask==2.3.2
Werkzeug<3.0
pytest==7.4.0
pytest-flask==1.2.0
backports.zoneinfo==0.2.1
